Fix issues with binomial_coef
binomial_coef had a few bugs. For example, 8 choose 8 would evaluate to 8, rather than 1.

That issue is now fixed, along with a bit of cleanup. I also changed the function to safely do computation on integral types without roundoff.

If you'd like to test--and to see the fixes in output--run:
  std:: cout << 
  gcem::binomial_coef(11, 0) << ' ' <<
  gcem::binomial_coef(11, 1) << ' ' <<
  gcem::binomial_coef(11, 2) << ' ' <<
  gcem::binomial_coef(11, 3) << ' ' <<
  gcem::binomial_coef(11, 4) << ' ' <<
  gcem::binomial_coef(11, 5) << ' ' <<
  gcem::binomial_coef(11, 6) << ' ' <<
  gcem::binomial_coef(11, 7) << ' ' <<
  gcem::binomial_coef(11, 8) << ' ' <<
  gcem::binomial_coef(11, 9) << ' ' <<
  gcem::binomial_coef(11, 10) << ' ' <<
  gcem::binomial_coef(11, 11) << std::endl;